Gate disabled-mode asserts on core class resolver generics
Drupal 11.x HEAD added a conditional return type to
getInstanceFromDefinition(), so the signature fallback narrows
class-string arguments natively there. Keep the version-stable
service-ID asserts unconditional and only gather the plain-object
class-string asserts when the core docblock has no @template.
