+++ tests/src/Type/data/drupal-class-resolver-disabled.php
@@ -4,21 +4,15 @@ namespace DrupalClassResolverDisabled;
 
 use Drupal\Core\DependencyInjection\ClassResolver;
 use Drupal\Core\DependencyInjection\ClassResolverInterface;
-use Drupal\service_map\MyService;
 use function PHPStan\Testing\assertType;
-
-class Foo {}
 
 function test(): void {
     // With classResolverReturnType disabled the extensions fall back to the
-    // declared signature types, so instanceof assertions remain meaningful.
-    assertType('object', (new ClassResolver())->getInstanceFromDefinition(Foo::class));
-    assertType('object', \Drupal::service('class_resolver')->getInstanceFromDefinition(Foo::class));
-    assertType('object', \Drupal::classResolver()->getInstanceFromDefinition(Foo::class));
-    assertType('object', \Drupal::classResolver(Foo::class));
+    // declared signature types. Service IDs are not class-strings, so these
+    // stay `object` even on cores whose docblocks declare generics.
     assertType('object', (new ClassResolver())->getInstanceFromDefinition('service_map.my_service'));
-    assertType('object', \Drupal::classResolver('service_map.my_service'));
-    assertType('object', (new ClassResolver())->getInstanceFromDefinition(MyService::class));
+    assertType('object', \Drupal::service('class_resolver')->getInstanceFromDefinition('service_map.my_service'));
+    assertType('object', \Drupal::classResolver()->getInstanceFromDefinition('service_map.my_service'));
 
     // The zero-argument call still resolves to the class resolver itself.
     assertType(ClassResolverInterface::class, \Drupal::classResolver());
